+++ program/rpi_pico_driver/src/parser/file.cpp
@@ -18,10 +18,10 @@ namespace parser {
         _payloadCount = 1;
         _payloadSize = _card->fileSize() - _primaryHeader.size();
     }
-    uint File::payloadSize(uint index) {
+    uint64_t File::payloadSize(uint32_t index) {
         return _payloadSize;
     }
-    bool File::getPayload(uint index, void *obj) {
+    bool File::getPayload(uint32_t index, void *obj) {
         auto *library = (Library *)obj;
         library->init(_card, _payloadOffset);
         return true;
@@ -30,7 +30,7 @@ namespace parser {
     // #### Public ####
     // Constructor & Destructor
     File::File() = default;
-    File::File(SDCard *card, uint offset) {
+    File::File(SDCard *card, uint64_t offset) {
         init(card, offset);
     }
 
@@ -43,13 +43,13 @@ namespace parser {
     }
 
     // Functions
-    void File::init(SDCard *card, uint offset) {
+    void File::init(SDCard *card, uint64_t offset) {
         _offset = offset;
         _card = card;
         readData();
         _size = _primaryHeader.size() + _payloadSize;
     }
-    uint File::payloadSize() {
+    uint64_t File::payloadSize() {
         return payloadSize(0);
     }
     Library File::getPayload() {
